+++ @navikt/aksel-icons/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @navikt/aksel-icons
 
+## 2.8.10
+
+## 2.8.9
+
 ## 2.8.8
 
 ### Patch Changes